Разделение js-файла на модули
